+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 
 ### Minor Changes
 
+## 3.5.6
+
+### Minor Changes
+
+- Added `MONAD` testnet.
+- Added `SETTLUS` testnet (Sepolia).
+
+## 3.5.5
+
+### Minor Changes
+
+- Added `SEI` mainnet.
+
 ## 3.5.4
 
 ### Major Changes